Release 13.0
Optimizing SISAL Compiler 13.0

+++ Backend/Library/AddStamp.c
@@ -11,10 +11,14 @@ void AddStamp( stamp, s )
      char stamp;
      char *s;
 { 
-  stamps[ stamp ] = s;
+  stamps[ (int)stamp ] = s;
 }
 
 /* $Log: AddStamp.c,v $
+ * Revision 1.2  1994/04/15  15:51:38  denton
+ * Added config.h to centralize machine specific header files.
+ * Fixed gcc warings.
+ *
  * Revision 1.1  1993/01/21  23:27:39  miller
  * Initial version of the IFX library.  It replaces the if[12]build.c
  * read.c timer.c util.c and write.c and if[12].h files from the
